Separate JS bindings from queue implementation
