+++ crates/javy/src/apis/timers/mod.rs
@@ -1,8 +1,7 @@
-use std::{
-    collections::BinaryHeap,
-    sync::{Arc, Mutex},
-    time::{SystemTime, UNIX_EPOCH},
-};
+use std::sync::{Arc, Mutex};
+
+mod queue;
+use queue::TimerQueue;
 
 use crate::{
     hold, hold_and_release,
@@ -112,109 +111,6 @@ impl TimersRuntime {
     }
 }
 
-/// Timer entry in the timer queue
-#[derive(Debug, Clone)]
-struct Timer {
-    id: u32,
-    fire_time: u64,           // milliseconds since UNIX epoch
-    callback: String,         // JavaScript code to execute
-    interval_ms: Option<u32>, // If Some(), this is a repeating timer
-}
-
-impl PartialEq for Timer {
-    fn eq(&self, other: &Self) -> bool {
-        self.fire_time == other.fire_time
-    }
-}
-
-impl Eq for Timer {}
-
-impl PartialOrd for Timer {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for Timer {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        // Reverse order for min-heap behavior
-        other.fire_time.cmp(&self.fire_time)
-    }
-}
-
-/// Global timer queue
-#[derive(Debug)]
-struct TimerQueue {
-    timers: BinaryHeap<Timer>,
-    next_id: u32,
-}
-
-impl TimerQueue {
-    fn new() -> Self {
-        Self {
-            timers: BinaryHeap::new(),
-            next_id: 1,
-        }
-    }
-
-    fn add_timer(
-        &mut self,
-        delay_ms: u32,
-        repeat: bool,
-        callback: String,
-        reuse_id: Option<u32>,
-    ) -> u32 {
-        let now = Self::now();
-
-        let id = reuse_id.unwrap_or_else(|| {
-            let id = self.next_id;
-            self.next_id += 1;
-            id
-        });
-
-        let timer = Timer {
-            id,
-            fire_time: now + delay_ms as u64,
-            callback,
-            interval_ms: if repeat { Some(delay_ms) } else { None },
-        };
-
-        self.timers.push(timer);
-        id
-    }
-
-    fn remove_timer(&mut self, timer_id: u32) -> bool {
-        let original_len = self.timers.len();
-        self.timers.retain(|timer| timer.id != timer_id);
-        self.timers.len() != original_len
-    }
-
-    fn get_expired_timers(&mut self) -> Vec<Timer> {
-        let now = Self::now();
-        let mut expired = Vec::new();
-        while let Some(timer) = self.timers.peek() {
-            if timer.fire_time <= now {
-                expired.push(self.timers.pop().unwrap());
-            } else {
-                break;
-            }
-        }
-
-        expired
-    }
-
-    fn has_pending_timers(&self) -> bool {
-        !self.timers.is_empty()
-    }
-
-    fn now() -> u64 {
-        SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap()
-            .as_millis() as u64
-    }
-}
-
 fn set_timeout<'js>(queue: &Arc<Mutex<TimerQueue>>, args: Args<'js>) -> Result<Value<'js>> {
     let (ctx, args) = args.release();
     let args = args.into_inner();
@@ -316,28 +212,6 @@ mod tests {
     use super::*;
     use crate::{Config, Runtime};
     use anyhow::Error;
-
-    #[test]
-    fn test_timer_queue() {
-        let mut queue = TimerQueue::new();
-
-        // Add some timers
-        let id1 = queue.add_timer(100, false, "console.log('timer1')".to_string(), None);
-        let id2 = queue.add_timer(50, false, "console.log('timer2')".to_string(), None);
-        let id3 = queue.add_timer(200, false, "console.log('timer3')".to_string(), None);
-
-        assert_eq!(id1, 1);
-        assert_eq!(id2, 2);
-        assert_eq!(id3, 3);
-
-        assert!(queue.has_pending_timers());
-
-        // Remove a timer
-        assert!(queue.remove_timer(id2));
-        assert!(!queue.remove_timer(999)); // Non-existent timer
-
-        assert!(queue.has_pending_timers());
-    }
 
     #[test]
     fn test_register() -> Result<()> {
